requirements.txt
caldav==1.3.9
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
icalendar==5.0.11
pytz==2024.1
requests==2.31.0
python-dateutil==2.8.2
